+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/unioncollege-webtech/puppies-vs-kittens",
   "dependencies": {
+    "body-parser": "^1.15.2",
     "express": "^4.13.3"
   }
 }
