+++ public/index.html
@@ -15,10 +15,10 @@
         <img src="cat.jpg" alt="Kitten" style="width:304px;height:228px;">
         <img src="dog.jpg" alt="Puppy" style="width:304px;height:228px;">
         <br>
-        <form method = "POST" action= "/kittens">
-            <button type="submit">Vote for Kittens!</button></form>
-        <form method = "POST" action= "/puppies">
-            <button type="submit">Vote for Puppies!</button></form>
+        <form method = "POST" action= "/vote4pet">
+            <button type="submit" value="kittens" name="choice">Vote for Kittens!</button></form>
+        <form method = "POST" action= "/vote4pet">
+            <button type="submit" value ="puppies" name="choice">Vote for Puppies!</button></form>
     </p>
 </body>
