+++ public/styles.css
@@ -45,3 +45,8 @@ h1{
 }
 
 form { display: inline; }
+
+
+body {
+	background-color: yellow;
+}
