+++ views/pugcss.css
@@ -1,0 +1,4 @@
+body {
+  color: blue; 
+  background-color:#6c7c7c;
+}
