+++ views/vote4pet.pug
@@ -1,0 +1,7 @@
+<!doctype html>
+html
+ head
+  title Puppies Vs. Kittens
+ body
+  h1 You voted for #{choice}.
+  p #{choice} has #{voteCount} vote(s).
